copy: Track worker queue size
Tracking the number of queued bytes per worker will allow optimizing the
number of in flight requests based on the actual requests size.

The goal is to allow large number of small requests, required to get
good performance, and in the same time limit the number of large
requests, that can be faster with lower number of requests.

Signed-off-by: Nir Soffer <nsoffer@redhat.com>