interop: Skip block-size tests on old qemu-nbd
The qemu-nbd on the CI's x86_64-ubuntu-1804 setup is so old it still
uses oldstyle negotiation, which lacks block size advertisements.
Skip in that case.

On x86_64-opensuse-tumbleweed, the CI timed out.  Obviously, the test
wedged (probably a mismatch where both nbdsh as client and old
qemu-nbd as server server are each waiting on the other to send more
data), but on timeout I don't have a nice .log file to see how far it
got before wedging.  Force qemu to quit after a reasonble time so that
I have more to debug by.

Fixes: 0a9adf6b ("api: Add set_request_block_size", v1.12.0)