interop: Skip test for qemu 3.1
The CI still complained for x86_64-debian-10, which appears to have
qemu 3.1 and reported minblock=512 even for '-f raw'.  That version of
qemu predated commit b0245d64 in v4.0 which flipped a switch to have
block sizes report as 1 when it makes sense.  Since v4.0 also
introduced 'qemu-nbd --list', we can write an easy 'requires' filter
for new-enough qemu-nbd.

Fixes: baae33e2