maint: Rename top level directory bash -> bash-completion
For testing uninstalled builds, libnbd's approach of a ./run wrapper
(which works with multiple in-tree binaries) is different than
nbdkit's approach of recommending the modification of $PATH (where a
single shim binary is all the more we needed).  Thus, we are less
likely to trip up on the '#!/usr/bin/env bash' problem of attempting
to execute our local 'bash' directory, which was addressed by nbdkit
commit b17fa77bc.  But maintaining consistent subdirectory names
between the two projects is still beneficial, even if we don't trip up
on the PATH modification pitfall.