docs: Fix POD verbatim paragraphs
As described in this nbdkit commit:

https://gitlab.com/nbdkit/nbdkit/-/commit/238da3442aab2ac226574e775ae8dd142d37402b

fix POD verbatim paragraphs throughout the documentation.

This change was made mechanically using the same Perl script as
appears in the above commit message.

I manually checked the POD fragments in generator/API.ml and although
there is a small amount of verbatim text, it is not affected.