python: Avoid memleak on (unlikely) module failure
Python 3.10 added PyModule_AddObjectRef() to more easily avoid a
common memory leak when adding to a module fails (unlikely in our case,
since we initialize early in the python process, but still something
we must worry about for corner-case correctness).  But since we target
older Python, we must check for errors and clean up ourselves.

Fixes: 259d46cb ("python: Raise a custom exception containing error string and errno.", v0.1.6)
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>