python: Improve doc comments for nbd.py
Enhance the documentation for nbd.Buffer.from_bytearray to match the
semantic changes from commit d477f7c7, before we change those
semantics yet again in an upcoming patch (copying is inefficient, we
are moving towards reusing the underlying buffer directly).

While touching this, PEP257 recommends that doc comments start with a
one line summary and be a complete sentence.  Tweak our existing
comments to comply.  It also recommends the use of """ strings, and
for u""" when a string contains Unicode; however, since it is easier
to write ' than " in OCaml pr statements, we stick with ''' and u'''
strings rather than following the PEP completely.

Reviewed-by: Richard W.M. Jones <rjones@redhat.com>