python: Plug uninit leak in nbd.Buffer.to_bytearray
When we created nbd.Buffer to handle both aio_pread and aio_pwrite, we
provided two different approaches to construct an instance.  The
constructor nbd.Buffer(int) allocates uninitialized data, to be filled
in by aio_pread (rather than wasting time pre-zeroing the data just to
be changed again - although that speedup is only possible when opting
in to h.set_pread_initialized(False)).  Meanwhile, the classmethod
nbd.Buffer.from_bytearray(...) allocates initialized data, to be
consumed by aio_pwrite.  However, this scheme makes it trivially easy
to use the wrong instance creation approach for a given handle
operation, and thereby leak uninitialized heap contents into the
destination with something as simple-looking as
h.aio_pwrite(nbd.Buffer(512), 0).  It doesn't help that while
bytearray(10) is all zeroes (a conscientious choice of the Python
language), and therefore nbd.Buffer.from_bytearray(10).is_zero() is
deterministically True, nbd.Buffer(10).is_zero() is non-deterministic;
nor that nbd.Buffer(10).to_bytearray() can allow Python code to play
with uninitialized data.

It is always a dangerous practice to leak heap contents; however, as
this leak is triggered only by poor client Python code, and not under
control of anything the server sends, we don't think it rises to the
level of the heap leak that was assigned CVE-2022-0485 as plugged back
in commit 8d444b41.

The solution employed here is to mark when a buffer has been
initialized, in nbd.Buffer.from_bytearray() and
h.aio_pread[_structured], as well as force-initialize an uninitialized
buffer before b.to_bytearray() or h.aio_pwrite.  Furthermore, we can
make b.is_zero() pretend an uninitialized buffer is all zeroes (since
the user can no longer easily get at any other contents).

The approach is not bulletproof - if h.aio_pread triggers a subsequent
read error in the C code (such as client-side bounds check or server
error response) AND the user has disabled h.set_pread_initialized(),
we've already marked the buffer initialized at that point even though
the buffer may not be fully written.  But that's okay - we've already
documented that a user opting in to that setting is responsible for
safe buffer practices, and should not dereference a read buffer when
an error is reported while retiring the command.

Generated code changes as follows:

| --- python/methods.h.bak	2022-06-06 07:43:33.167181594 -0500
| +++ python/methods.h	2022-06-06 07:43:40.494190038 -0500
| @@ -31,6 +31,7 @@
|  struct py_aio_buffer {
|    Py_ssize_t len;
|    void *data;
| +  bool initialized;
|  };
|
|  extern char **nbd_internal_py_get_string_list (PyObject *);
| --- python/methods.c.bak	2022-06-06 07:43:33.165181592 -0500
| +++ python/methods.c	2022-06-06 07:56:25.109094074 -0500
| @@ -3201,6 +3201,7 @@ nbd_internal_py_aio_pread (PyObject *sel
|    completion_user_data->buf = buf;
|    offset_u64 = offset;
|
| +  buf_buf->initialized = true;
|    ret = nbd_aio_pread (h, buf_buf->data, buf_buf->len, offset_u64, completion, flags_u32);
|    completion_user_data = NULL;
|    if (ret == -1) {
| @@ -3274,6 +3275,7 @@ nbd_internal_py_aio_pread_structured (Py
|    Py_INCREF (py_chunk_fn);
|    chunk_user_data->fn = py_chunk_fn;
|
| +  buf_buf->initialized = true;
|    ret = nbd_aio_pread_structured (h, buf_buf->data, buf_buf->len, offset_u64, chunk, completion, flags_u32);
|    chunk_user_data = NULL;
|    completion_user_data = NULL;
| @@ -3335,6 +3337,10 @@ nbd_internal_py_aio_pwrite (PyObject *se
|    completion_user_data->buf = buf;
|    offset_u64 = offset;
|
| +  if (!buf_buf->initialized) {
| +    memset (buf_buf->data, 0, buf_buf->len);
| +    buf_buf->initialized = true;
| +  }
|    ret = nbd_aio_pwrite (h, buf_buf->data, buf_buf->len, offset_u64, completion, flags_u32);
|    completion_user_data = NULL;
|    if (ret == -1) {
| --- python/nbd.py.bak	2022-06-06 07:43:33.169181596 -0500
| +++ python/nbd.py	2022-06-06 07:43:40.516190063 -0500
| @@ -150,7 +150,9 @@ class Buffer(object):
|      def is_zero(self, offset=0, size=-1):
|          '''Returns true if and only if all bytes in the buffer are zeroes.
|
| -        Note that a freshly allocated buffer is uninitialized, not zero.
| +        Note that although a freshly allocated buffer is uninitialized,
| +        this will report it as all zeroes, as it will be force-initialized
| +        to zero before any code that can access the buffer's contents.
|
|          By default this tests the whole buffer, but you can restrict
|          the test to a sub-range of the buffer using the optional

Fixes: f05b8ec5 ("python: Change aio_buffer into nbd.Buffer class.", v0.9.8)
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>