python: Enhance tests of nbd.Buffer
Add some more coverage of existing behavior, so we can better track
that we are not introducing unintended changes in upcoming patches.

While at it, compress the computation of the expected output from the
nbdkit pattern plugin by utilizing Python's ability to cast an
array.array of longs to a correctly-byte-ordered byte-wise memoryview.

Reviewed-by: Richard W.M. Jones <rjones@redhat.com>