python: Slice structured read callback buffer from original
The Py_BuildValue "y#" format copies data; this is because Python
assumes our C memory can go out of scope, while the user's python
callback can stash off whatever bytearray object it got.  But this
copying is inefficient.  Now that we already have a Python buffer-like
object in scope for the duration of all structured reads
(pread_structured since commit f8c76c01, aio_pread_structured since
commit 4f15d0e9), we know that the C memory parameter to our
chunk_callback is already (a portion of) the underlying contiguous
memory of an original Python object.  If we pass the Python callback a
memoryview object over the correct slice of that memory, then Python
will correctly handle the intricacies of what happens if the callback
saves our memoryview object for later use, all without any data
copying overhead, making use of the pread_structured form much faster.

We basically want to code up the C equivalent to the python expression
'view = memoryview(buf).toreadonly()[start:end]', where start and end
are determined by comparing subbuf/count against the original buf.
But in C, that involves creating PyObject integers to form a PySlice,
then using PyObject_GetItem to apply the slice to a PyMemoryView (I
don't know of any faster tricks; googling does not find any quick
hits).  Extract this work into a new utils.c helper function, to avoid
repetition (and in case we figure out a way to do it in fewer lines by
directly manipulating Py_buffer objects).

The testsuite is updated to cover examples of the data persistence
aspect, demonstrating that our stashed slice is still tied to the
returned Python buffer, even though the callback is long since
completed (prior to this patch, stash was an independent bytes copy,
and while it persisted on after the callback, it is frozen in time to
the contents of the buffer at the time the callback was reached).  The
fact that data is now shared is is an API change, but one unlikely to
cause problems - it is very unusual for a callback function to try and
stash sub-buffers for later use (neither libnbd nor nbdkit was
utilizing that in their testsuites before this patch).

To demonstrate the speedup now that there is less copying involved per
callback, I tested:

$ export script='
def f(b,o,s,e):
 pass
m=1024*1024
size=h.get_size()
h.set_pread_initialize(False)
for i in range(size // m):
 buf = h.pread_structured(m, m*i, f)
 buf = None
'
$ time ./run nbdkit -U - memory 10G --run 'nbdsh -u "$uri" -c "$script"'

On my machine, this took 16.9s with libnbd 1.12.3, 15.5s pre-patch,
and 3.6s post-patch, for more than 4x speedup in the best case.  Much
of the differences boil down to how efficient the Python garbage
collector is able to spot liveness of various buffers, to reuse rather
than allocate new buffers on the Python side; removing the 'buf =
None' line speeds up libnbd 1.12.3 to 4.5s (better reuse of the "y#"
memory in the callback when backed by C memory), 9.0s pre-patch
(copying "y#" from Python memory has some negative caching effects),
and 3.0s post-patch (zero-copying is always best).

A variation with aio_pread_structured reusing the same nbd.Buffer is
not quite as drastic: 3.9s with libnbd 1.12.3, 4.0s pre-patch, and
3.4s with this patch, but still shows that it is beneficial.

The corresponding diff to generated code is:

| --- python/methods.h.bak	2022-06-10 13:56:13.887618074 -0500
| +++ python/methods.h	2022-06-10 13:56:20.850628290 -0500
| @@ -36,6 +36,7 @@
|  extern int nbd_internal_py_init_aio_buffer (PyObject *);
|  extern PyObject *nbd_internal_py_get_nbd_buffer_type (void);
|  extern PyObject *nbd_internal_py_wrap_errptr (int);
| +extern PyObject *nbd_internal_py_get_subview (PyObject *, const char *, size_t);
|
|  static inline struct nbd_handle *
|  get_handle (PyObject *obj)
| --- python/methods.c.bak	2022-06-10 13:56:13.882618066 -0500
| +++ python/methods.c	2022-06-10 14:03:07.059233226 -0500
| @@ -75,13 +75,15 @@ chunk_wrapper (void *user_data, const vo
|
|    PyGILState_STATE py_save = PyGILState_Ensure ();
|    PyObject *py_args, *py_ret;
| +  PyObject *py_subbuf = NULL;
|    PyObject *py_error = NULL;
|
| +  py_subbuf = nbd_internal_py_get_subview (data->view, subbuf, count);
| +  if (!py_subbuf) { PyErr_PrintEx (0); goto out; }
|    py_error = nbd_internal_py_wrap_errptr (*error);
|    if (!py_error) { PyErr_PrintEx (0); goto out; }
|
| -  py_args = Py_BuildValue ("(y#KIO)", subbuf, (int) count, offset, status,
| -                           py_error);
| +  py_args = Py_BuildValue ("(OKIO)", py_subbuf, offset, status, py_error);
|    if (!py_args) { PyErr_PrintEx (0); goto out; }
|
|    py_ret = PyObject_CallObject (data->fn, py_args);
| @@ -106,6 +108,7 @@ chunk_wrapper (void *user_data, const vo
|    };
|
|   out:
| +  Py_XDECREF (py_subbuf);
|    if (py_error) {
|      PyObject *py_error_ret = PyObject_GetAttrString (py_error, "value");
|      *error = PyLong_AsLong (py_error_ret);
| @@ -2367,6 +2370,8 @@ nbd_internal_py_pread_structured (PyObje
|    /* Increment refcount since pointer may be saved by libnbd. */
|    Py_INCREF (py_chunk_fn);
|    chunk_user_data->fn = py_chunk_fn;
| +  chunk_user_data->view = nbd_internal_py_get_aio_view (buf, PyBUF_WRITE);
| +  if (!chunk_user_data->view) goto out;
|
|    Py_BEGIN_ALLOW_THREADS
|    ret = nbd_pread_structured (h, PyByteArray_AS_STRING (buf), count,
| @@ -3298,6 +3303,8 @@ nbd_internal_py_aio_pread_structured (Py
|    /* Increment refcount since pointer may be saved by libnbd. */
|    Py_INCREF (py_chunk_fn);
|    chunk_user_data->fn = py_chunk_fn;
| +  chunk_user_data->view = nbd_internal_py_get_aio_view (buf, PyBUF_WRITE);
| +  if (!chunk_user_data->view) goto out;
|
|    if (nbd_internal_py_init_aio_buffer (buf) < 0) goto out;
|    Py_BEGIN_ALLOW_THREADS

Acked-by: Richard W.M. Jones <rjones@redhat.com>