lib/uri.c: Don't corrupt memory on error path
We incremented list++ instead of i++ resulting in potential memory
corruption if the error path is taken when splitting URI query
strings.

Luckily this is very hard to trigger.  xmlURIUnescapeString doesn't
actually check for URI validity (only malloc failure), and the only
other way to reach this code is via strdup("") returning NULL which is
very unlikely.

Found by Coverity:

  Error: ARRAY_VS_SINGLETON (CWE-119): [#def51]
  libnbd-1.12.5/lib/uri.c:169: address_of: Taking address with "&queries" yields a singleton pointer.
  libnbd-1.12.5/lib/uri.c:169: callee_ptr_arith: Passing "&queries" to function "parse_uri_queries" which uses it as an array. This might corrupt or misinterpret adjacent memory locations.
  #  167|       goto cleanup;
  #  168|     }
  #  169|->   if (parse_uri_queries (uri->query_raw, &queries) == -1) {
  #  170|       set_error (EINVAL, "unable to parse URI queries: %s", uri->query_raw);
  #  171|       goto cleanup;

Fixes: commit 6f3eee2ede49bb251c67f8818665757227a5e41d