build: Remove use of $^
FreeBSD's default make doesn't support $^ which is apparently a GNU
make extension.  This leads to weird errors such as the generator
being built but empty.  Although you're supposed to be using GNU make,
this particular issue causes very weird and hard to diagnose problems
with the build, so replace use of $^ throughout.