states: Fix double-translation of structured reply errors
Due to poor merge resolution of several patch series that were in
flight at the same time, commit 2c93dc5a reintroduces the very same
category of bug that commit 71d1cd17 tried to fix in relation to
commit 97fab5e9.  On Linux, where protocol NBD_E* and errno values E*
are identical, the bug cannot be observed (the second call is
idempotent).  But on other platforms, a second call to to translate
what is a system errno value instead of a wire value may end up
slamming the errno value to EINVAL instead of what was transmitted
over the wire.

Fixes: 2c93dc5a ("states: Wire in a read callback", v0.1.5)