generator: Use static buffer for discarding reads
No need to malloc()/free() a throw-away buffer, when we can just use a
static one.  We normally don't even have to worry about thread-safety
issues between separate connections (as long as we never read the
buffer, we don't care if separate nbd handles clobber the same buffer
at the same time); but for DUMP_PACKETS, we have to use stack
allocations instead of static allocation (this makes stack utilization
larger and may fragment into more reads, but this is okay for an
expensive debugging mode that isn't even going to be compiled in
normal practice).

Message-Id: <20220812020638.856332-2-eblake@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>