states: Refactor where interrupted reply state is stored
For a given nbd handle, there can be at most one active reply at a
time.  Storing a resume state per command is wasteful compared to
storing a single resume state in the overall handle.  Furthermore, if
we want to be able to skip bogus server replies to an unexpected
cookie, we need a way to resume reading bytes off the wire even when
there is no matching h->reply_cmd for the current cookie.