states: Skip server replies to unknown command cookies
Instead of killing the connection when the server gives us an
unexpected structured or error reply for a cookie we don't recognize,
we can log the issue and otherwise ignore that packet (an unexpected
successful simple reply is harder if we did not negotiate structured
replies; see the lengthy comment in states-reply-simple.c).  A
compliant server will never do this to us, but a transmission error
might.  Here's a quick patch to nbdkit to demonstrate such a scenario:

| diff --git i/server/protocol.c w/server/protocol.c
| index 2ac77055..55b27e2e 100644
| --- i/server/protocol.c
| +++ w/server/protocol.c
| @@ -742,7 +742,7 @@ protocol_recv_request_send_reply (void)
|        (cmd == NBD_CMD_READ || cmd == NBD_CMD_BLOCK_STATUS)) {
|      if (!error) {
|        if (cmd == NBD_CMD_READ)
| -        return send_structured_reply_read (request.handle, cmd,
| +        return send_structured_reply_read (0x100^request.handle, cmd,
|                                             buf, count, offset);
|        else /* NBD_CMD_BLOCK_STATUS */
|          return send_structured_reply_block_status (request.handle,

Note that if a server bit-flips a cookie, our original command waiting
for the right cookie may never get a completion packet from the
server; thus, testing this requires aio commands (unless we someday
add APIs to make it easier for blocking commands that give up when
they hit a timeout).  With the following command line and the hack to
nbdkit above:

$ ./run /path/to/nbdkit -U - memory 1M --run 'nbdsh -u "$uri" -c "\
buf=nbd.Buffer(1)
cookie=h.aio_pread(buf, 0)
try:
  print(h.poll(1))
except nbd.Error as ex:
  print(ex.string)
h.flush()
print(h.get_size())
"'

pre-patch we see that the connection is killed when h.poll()
encounters the garbage reply, before we get to the second command:

nbd_poll: no matching cookie found for server reply, this is probably a bug in the server
nbdkit: memory.1: error: read request: Connection reset by peer
nbdsh: command line script failed: nbd_flush: invalid state: DEAD: the handle must be connected with the server: Invalid argument

while post-patch the read never retires, but the unexpected cookie is
skipped and we get to the flush:

1
1048576

Watching the LIBNBD_DEBUG=1 stream, we also see:

libnbd: debug: nbdsh: nbd_poll: skipped reply for unexpected cookie 281474976710657, this is probably a bug in the server

We may decide to give libnbd a strictness knob to control whether we
ignore such commands by default, or move to DEAD as soon as we detect
a server bug, but that would be a later patch.