copy: In tests, ensure pwrite scripts consume stdin
These tests failed on macOS with this error:

  nbdkit: eval[1]: error: /tmp/nbdkit8bDHNo/pwrite: write: Broken pipe
  nbdkit: eval[1]: error: /tmp/nbdkit8bDHNo/pwrite: script exited with error, but did not print an error message on stderr

A simplified example is:

  ./nbdkit -fv eval get_size='echo 7E' pwrite='exit 0' \
          --run 'nbdsh -u $uri -c "h.pwrite(bytearray(33554432),0)"'

The problem is that the pwrite script does not consume stdin so nbdkit
gets an EPIPE error.  This is a problem in the test, not nbdkit.

(cherry picked from commit c713529e9fd0641b2d73f764517b5f9c21a767fd)