ublk: Add new nbdublk program
ublk is a facility in Linux 6.0 which allows block devices to be
written in userspace.  This patch contains a new program that
implements an NBD client using this facility.  It is roughly the
equivalent of the existing nbd.ko kernel module, except of course
implemented as a userspace process.

This requires libublksrv (https://github.com/ming1/ubdsrv),
liburing >= 2.2, and of course a sufficiently recent Linux kernel.

Example usage:

  # modprobe ublk_drv
  # nbdublk /dev/ublkb0 nbd://server
  # mount /dev/ublkb0 /mnt
  ...
  # umount /mnt
  # ublk list
  # ublk del -n 0

As well as NBD URIs pointing to TCP/IP servers (like nbd://server),
all other features of libnbd are available, such as TLS, Unix domain
sockets, vsock, "[ NBD-SERVER ARGS ... ]" &c.

Thanks Ming Lei who advised on and authored much of this code.