api: Allow nbd_opt_list_meta_context() without SR
Upstream NBD clarified (see NBD commit 13a4e33a8) that since
NBD_OPT_LIST_META_CONTEXT is stateless on the server side, it is
acceptable (but not mandatory) for servers to accept it without the
client having pre-negotiated structured replies.  We aren't quite
stateless on the client side yet - that will be fixed in a later patch
to keep this one small and easier to test.  The test is in the next
patch, to make it easier to temporarily reorder the series to see that
it catches the problem.

Acked-by: Richard W.M. Jones <rjones@redhat.com>