api: Make nbd_opt_list_meta_context stateless
Since NBD_OPT_LIST_META_CONTEXTS is stateless from the server's point
of view, there is no reason to make it clear state on the client's
side.  A previous patch already fixed it to not wipe h->meta_contexts
for NBD_OPT_LIST_META_CONTEXTS (since that tracks the server state
affected only by NBD_OPT_SET_META_CONTEXTS); but we were still wiping
h->exportsize and friends.  Better is to wipe size information only if
we actually enter NEWSTYLE.OPT_GO.START to attempt NBD_OPT_GO or
NBD_OPT_INFO.  This change will also help when a future patch adds
nbd_opt_set_meta_context() as another API that wants to modify
h->meta_contexts but not h->exportsize.

Testsuite coverage for this is in the next patch, for ease of review
(that is, this is one case where it is easy to swap the patch order to
see a failure fixed by this patch).

Acked-by: Richard W.M. Jones <rjones@redhat.com>