info: Explicitly skip NBD_OPT_SET_META_CONTEXT in --list mode
When listing information about an export, we do not use
nbd_block_status() or nbd_can_meta_context(), because we were already
utilizing the callback to nbd_opt_list_meta_context() to collect
supported context names.  In practice, because we also never called
nbd_add_meta_context() (important because nbd_opt_list_meta_context()
only outputs all possible contexts when passed an empty list on
input), we were not causing nbd_opt_info() to invoke
NBD_OPT_SET_META_CONTEXT (SET on an empty list selects nothing, so the
state machine short-circuits that step).  However, now that we have
the API to express our intentions, we might as well be explicit that
we don't need the server to set any contexts, even if such a change
does not impact the amount of traffic sent over the wire in list mode.

Acked-by: Laszlo Ersek <lersek@redhat.com>
Acked-by: Richard W.M. Jones <rjones@redhat.com>