ocaml: More idiomatic use of assert in tests
It is bad form to use an expression with side effects inside an
assert, as the exprssion remains untested if compiled with -noassert.
While the language documents[1] that it special-cases 'assert false'
to 'raise (Assert_failure ...)', it is not obvious whether this
syntactic sugar also occurs for 'assert (false)'.  And in general,
avoiding spurious () when an expression must be bool is useful.

[1] https://v2.ocaml.org/manual/expr.html

(cherry picked from commit 1c33db947fb3d7166e04dd691649aff05b2bf6fa)