generator: Add RUInt64 for 64-bit counters without error
The next patch will add some APIs that expose long-term statistics
counters.  A counter is always accessible (no need to return -1; if
the handle is new the count is still 0, and even after the handle is
in DEAD the counter is still useful).  But it is also long-running;
the existing RUInt might only be 32 bits, while it is easy to prove
some counters (bytes transmitted, for example) will need 64-bits.  So
time to plumb in a new return type.

Using signed int64 in OCaml bindings is okay (actually sending 2^63
bytes of data to overflow the counter takes a LOOONG time, so we don't
worry if we can't get to 2^64).