api: Add new nbd_stats_*() statistics counters
Expose our internal counters via a family of new APIs:
nbd_stats_{bytes,chunks}_{sent,received}().

I was trying to write a test that determined whether a failure was due
to a short-circuit test on the client side or an actual failure
returned by the server.  Such a test is a lot easier if we can watch
status counters increase according to how much traffic goes over the
wire.  We may add more stats later (for example, how many bytes were
written by nbd_[aio_]pread/write, or even how many times we had to
grab the lock as part of an nbd_* command); or even to reset the
counters for determining the protocol overhead around a particular set
of I/O options, but this is a fun start.

In this patch, update an example to serve as a compile time test of
the new API; the next patch will add a runtime test of some simple
uses while also checking the language bindings.

We intentionally do not document exactly when chunks increment; this
is to allow us to change the implementation should it be necessary,
without breaking the API (basically, applications can track that
counters increase when the server was involved, but should not expect
a particular number of bytes/chunks after a given API call).