tests: Add language binding tests for stats collection
This proves that the stat counters increment as desired, as well as
proving that our RUInt32 generator type works.  The test is slightly
fragile to whether nbdkit chooses to reply with simple or structured
replies when structured replies are negotiated.

This commit is also a showcase of whether I can do 64-bit math in
various languages (C's terseness in 'a == b + (c > d)' is annoying to
replicate in languages that don't like playing fast and loose with
types).  :)